requirements.txt
#--- Backend ---
faiss-cpu==1.7.4
langchain==0.0.188
requests==2.31.0
numpy==1.24.3
tqdm==4.66.1
python-dotenv==1.0.0

beautifulsoup4==4.12.2
lxml==4.9.3

fastapi==0.95.2
uvicorn==0.24.0
